fix: accept numbers as keys for ngx.shared.DICT methods
This updates the annotations on the ngx.shared.DICT methods to accept
numbers as keys in addition to strings.

Technically speaking, all of the shdict methods implemented in
[resty.core.shdict][1] will accept any non-nil type as a key so long as
it can be coerced to a string with `tostring()` (yes, even tables,
booleans, and userdata like `ngx.null`). However, this does not include
_all_ shdict methods. Notably, list methods like `lpush()`/`lpop()` will
raise an error if non-string, non-number values are used as keys.

While I'd argue that using floats as keys is a bad/unsafe practice
(integers should be just fine), I don't really want to split hairs any
further on this.

[1]: https://github.com/openresty/lua-resty-core/blob/31fae862a1ed64033591f991fadb0dd80358ba0b/lib/resty/core/shdict.lua